split mysqlVisitor
