+++ shardingsphere-sql-parser/shardingsphere-sql-parser-engine/src/test/java/org/apache/shardingsphere/sql/parser/core/rule/fixture/TestParserEntry.java
@@ -40,7 +40,7 @@ public final class TestParserEntry implements SQLParserEntry {
     }
     
     @Override
-    public Class<? extends SQLVisitor> getVisitorClass() {
+    public Class<? extends SQLVisitor> getVisitorClass(final String visitorName) {
         return null;
     }
 }
