+++ shardingsphere-sql-parser/shardingsphere-sql-parser-mysql/src/main/java/org/apache/shardingsphere/sql/parser/MySQLParserEntry.java
@@ -47,7 +47,10 @@ public final class MySQLParserEntry implements SQLParserEntry {
     }
     
     @Override
-    public Class<? extends SQLVisitor> getVisitorClass() {
+    public Class<? extends SQLVisitor> getVisitorClass(final String visitorName) {
+        if (MySQLDMLVisitor.class.getSimpleName().contains(visitorName)) {
+            return MySQLDMLVisitor.class;
+        }
         return MySQLVisitor.class;
     }
 }
