+++ shardingsphere-sql-parser/shardingsphere-sql-parser-oracle/src/main/java/org/apache/shardingsphere/sql/parser/OracleParserEntry.java
@@ -47,7 +47,7 @@ public final class OracleParserEntry implements SQLParserEntry {
     }
 
     @Override
-    public Class<? extends SQLVisitor> getVisitorClass() {
+    public Class<? extends SQLVisitor> getVisitorClass(final String visitorName) {
         return OracleVisitor.class;
     }
 }
