+++ shardingsphere-sql-parser/shardingsphere-sql-parser-spi/src/main/java/org/apache/shardingsphere/sql/parser/spi/SQLParserEntry.java
@@ -53,7 +53,8 @@ public interface SQLParserEntry {
     /**
      * Get SQL visitor class type.
      *
+     * @param visitorName visitor name
      * @return SQL visitor class type
      */
-    Class<? extends SQLVisitor> getVisitorClass();
+    Class<? extends SQLVisitor> getVisitorClass(final String visitorName);
 }
